refactor: Update imports and context usage in property controls to fix cyclic dependency
