+++ app/client/src/components/propertyControls/InputTextControl.tsx
@@ -1,9 +1,6 @@
-import React from "react";
-import type { ControlProps } from "./BaseControl";
-import BaseControl from "./BaseControl";
-import { StyledDynamicInput } from "./StyledControls";
 import type { InputType } from "components/constants";
 import type { CodeEditorExpected } from "components/editorComponents/CodeEditor";
+import { slashCommandHintHelper } from "components/editorComponents/CodeEditor/commandsHelper";
 import type { FieldEntityInformation } from "components/editorComponents/CodeEditor/EditorConfig";
 import {
   CodeEditorBorder,
@@ -12,11 +9,14 @@ import {
   EditorTheme,
   TabBehaviour,
 } from "components/editorComponents/CodeEditor/EditorConfig";
-import { CollapseContext } from "pages/Editor/PropertyPane/PropertySection";
-import LazyCodeEditor from "../editorComponents/LazyCodeEditor";
-import type { AdditionalDynamicDataTree } from "utils/autocomplete/customTreeTypeDefCreator";
 import { bindingHintHelper } from "components/editorComponents/CodeEditor/hintHelpers";
-import { slashCommandHintHelper } from "components/editorComponents/CodeEditor/commandsHelper";
+import { CollapseContext } from "pages/Editor/PropertyPane/PropertyPaneContexts";
+import React from "react";
+import type { AdditionalDynamicDataTree } from "utils/autocomplete/customTreeTypeDefCreator";
+import LazyCodeEditor from "../editorComponents/LazyCodeEditor";
+import type { ControlProps } from "./BaseControl";
+import BaseControl from "./BaseControl";
+import { StyledDynamicInput } from "./StyledControls";
 
 export function InputText(props: {
   label: string;
